Update admin_groups.php
added this on line 53 and 54:
'Mod promote users label'		=>	'Allow moderators to promote users',
'Mod promote users help'		=>	'If moderator privileges are enabled, allow users in this group to promote users.',
and translated. Please check if the danish is updated with proper lines compare to the english.. also danish translation seems to contain 4 more files then the english?